pyproject.toml
[project]
name = "control-inventario-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "django>=5.2.8",
    "django-cors-headers>=4.9.0",
    "django-ninja>=1.5.0",
    "dotenv>=0.9.9",
]
